Switch again to al-folio
Seems to be better maintained, which will make life easier. I tried to
configure the previous template, but ran into a number of issues. I
could work around them, but it seemed like it was going to be a big
pain point, so better to just shift now. I can always switch back
later if I decide it makes sense then